% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gauss_rec_to_em_input.R
\name{add_gauss_rec_to_em_input}
\alias{add_gauss_rec_to_em_input}
\title{Add Gaussian environmental recruitment settings to EM input}
\usage{
add_gauss_rec_to_em_input(em_input, gauss_rec_em)
}
\arguments{
\item{em_input}{List. Assessment model input object.}

\item{gauss_rec_em}{List specifying the Gaussian recruitment effect.}
}
\value{
Modified \code{em_input} object.
}
\description{
Adds an optional Gaussian environmental effect on recruitment to a WHAM-style
EM input object. This function modifies both \code{data}, \code{par}, and
\code{map} components so the resulting object can be passed to model fitting.
}
\details{
If \code{gauss_rec_em$use = FALSE}, placeholder values are inserted and the
corresponding parameters are fixed via \code{map}.
}
